+++ cfme/rest/__init__.py
@@ -26,3 +26,27 @@ def service_catalogs(request, rest_api):
             rest_api.collections.service_catalogs.action.delete(*delete_scls)
 
     return scls
+
+
+def categories(request, rest_api, num=1):
+    ctg_data = [{
+        'name': 'test_category_{}_{}'.format(fauxfactory.gen_alphanumeric().lower(), _index),
+        'description': 'test_category_{}_{}'.format(fauxfactory.gen_alphanumeric().lower(), _index)
+    } for _index in range(0, num)]
+    ctgs = rest_api.collections.categories.action.create(*ctg_data)
+    for ctg in ctgs:
+        wait_for(
+            lambda: rest_api.collections.categories.find_by(description=ctg.description),
+            num_sec=180,
+            delay=10,
+        )
+
+    @request.addfinalizer
+    def _finished():
+        ids = [ctg.id for ctg in ctgs]
+        delete_ctgs = [ctg for ctg in rest_api.collections.categories
+            if ctg.id in ids]
+        if len(delete_ctgs) != 0:
+            rest_api.collections.categories.action.delete(*delete_ctgs)
+
+    return ctgs
